+++ src/cqspoolterrain/pipe/straight.py
@@ -23,6 +23,11 @@ def straight(
         length = 75, 
         connector_length=2, 
         connector_radius = 11.5,
+        render_hollow=True,
+        hollow_padding = 6,
+        hollow_radius_padding = 3,
+        render_through_hole = True,
+        through_radius = 5,
         debug_magnets = False
     ):
     connector_plate = cq.Workplane("XY").cylinder(connector_length, connector_radius).rotate((0,1,0),(0,0,0),90).translate((0,0,connector_radius+.5))
@@ -47,5 +52,29 @@ def straight(
         z_lift = 6,
         debug = debug_magnets
     )
+    
+    if render_hollow:
+        center_cut = (
+            cq.Workplane("XY")
+            .cylinder(
+                length-hollow_padding,
+                connector_radius-hollow_radius_padding
+            )
+            .rotate((0,1,0),(0,0,0),90)
+            .translate((0,0,connector_radius))
+        )
+        barrier_plates_magnets = barrier_plates_magnets.cut(center_cut)
+        
+    if render_through_hole:
+        through_cut = (
+            cq.Workplane("XY")
+            .cylinder(
+                length,
+                through_radius
+            )
+            .rotate((0,1,0),(0,0,0),90)
+            .translate((0,0,connector_radius))
+        )
+        barrier_plates_magnets = barrier_plates_magnets.cut(through_cut)
 
     return barrier_plates_magnets