+++ changes.md
@@ -16,6 +16,7 @@
 * Added pipe.platform
 * Update the cqindustry version to 0.0.8
   * Brings in newest features from cqdqueryhelper 0.1.6 and cqterrain 0.2.0
+* Added hollow cut out option to pipe.straight
 
 ## 0.0.5
 * Added ladders to the powerStation.